Updated landing page and styling
